proyecto-retos-ia-backend/requirements.txt
Flask>=2.2
Flask-Cors>=3.0
Flask-SQLAlchemy>=3.0
psycopg2-binary>=2.9
python-dotenv>=0.21
